Remove generated files from the repository and add them to .gitignore
This PR is meant to start a discussion, and I won't be mad if yall think
it's a bad idea.

Generally it's not advisable to have generated files in your repository,
the main reason being that just building the project makes your
workspace dirty. Plus, it's easy to get merge conflicts in generated
files when realisitcly nobody cares; it just gets genearted at code
generation time.

It's also possible that I missed some--namely I wasn't sure if `main.c`
had any non-generated code in it. If it doesn't I can add it.